+++ src_ts/components/pages/interventions/intervention-tab-pages/intervention-details/document-details/selectors.ts
@@ -1,23 +1,18 @@
 import {createSelector} from 'reselect';
 import {Intervention} from '../../common/models/intervention-types';
-
-export const currentIntervention = (state: any) => state.interventions.current;
+import {DocumentDetails, DocumentDetailsPermissions} from './documentDetailsModels';
+import {currentIntervention, currentInterventionPermissions} from '../../selectors';
 
 export const selectDocumentDetails = createSelector(currentIntervention, (intervention: Intervention) => {
-  return {
-    details: {
-      title: intervention.title,
-      context: intervention.context,
-      implementation_strategy: intervention.implementation_strategy,
-      ip_progr_contrib: intervention.ip_progr_contrib
-    },
-    permissions: {
-      edit: {
-        title: true,
-        context: true,
-        implementation_strategy: true,
-        ip_progr_contrib: true
-      }
-    }
-  };
+  return new DocumentDetails(intervention);
 });
+
+export const selectDocumentDetailsPermissions = createSelector(
+  currentInterventionPermissions,
+  (permissions: Intervention) => {
+    return {
+      edit: new DocumentDetailsPermissions(permissions!.edit),
+      required: new DocumentDetailsPermissions(permissions!.required)
+    };
+  }
+);
